Inner the main() method was created an array of figures. The first half of figures in this array are generated with random parameters.
For this purpose was created two more classes:

ColorSupplier with public String getRandomColor() method - for generating random color,
and FigureSupplier with public Figure getRandomFigure() method - for generating figure with random properties.

The other half of figures are the same, default parameters.

For this purpose was create a new method in the FigureSupplier class:

public Figure getDefaultFigure();

After generating the array, we can display the entire list of figures.