Added a sub-class of the Figure class, which is inherited from it. The class has a constant with its own name and unique parameter fields. Added constructor for these fields.
The overridden getInfo() method, inside which sout is called, contains all the information about the current object.

The overridden getArea() method is called from the getInfo() method.
Inside the area of the figure is calculated and the answer is returned in the form of a double.

The received response is formatted using the format() method, based on the constant of the DecimalFormat variable in the GetArea() interface.