Added an empty Figure array, with a length of 6. The FigureSupplier object was initialized before the loop.
Added loop, within which the getRandomFigure method is called to generate a new figure in the i-th element of the array.
As soon as int i exceeds half the length of the figures array, the getDefaultFigure() method is called, which adds the default figure to the remaining half of the array.

After that, another loop is added that processes the entire figures array, and displays information about all stored elements using the getInfo () method.