Created Classes for Every figure. Created an abstract class Figure (I could also use interface Class for getArea, toDraw. But I realized that all the figures will use those methods, so I stick to the abstract class)
ColorSupplier and FigureSupplier classes were created.
In FigureSuplier I've created all the static valuables that were required, used switch to randomize Figures.

in Main I've created an array of Figures and manually declared positions in an array (I could've used for loop for that, if needed). and with for loop I've executed toDraw method.