Create classes for various shapes, generate random and default figures
Create classes for different shapes such as Square, Rectangle, RightTriangle, Circle, and IsoscelesTrapezoid. Each class has a color field and some unique properties such as radius for Circle, firstLeg and secondLeg for RightTriangle, etc. Also, all the classes have the getArea() and draw() methods for calculating the area and drawing the figure, respectively.

To generate random figures, two more classes ColorSupplier and FigureSupplier have been created. ColorSupplier has a getRandomColor() method for generating random colors, and FigureSupplier has a getRandomFigure() method for generating a random figure with random properties.

To generate default figures, a new method getDefaultFigure() has been added to the FigureSupplier class. This method always returns a white Circle with a radius of 10.

In the main() method, an array of figures is created with random and default figures. The entire list of objects is then displayed using the draw() method.