Update Figure.java
use drawFigure() like abstract method and FigureSupplier and AreaCalculator like Interface.
but I didn’t really understand. If all figures definitely have a drawFigure method, can it be made an abstract method? It’s just that in the description this method was described as behavior, and behavior is interfaces, how to determine where to place the method? The same question with the AreaCalculator method, which I first made as an abstract method, but according to tests it was necessary to add 1 more interface, so I remade it as an interface