First approach to the OPP advanced task.
I have a problem with test.
StructureTest.init:27 You should not rename base core.basesyntax package and path to project and project name should not contain spaces or words in cyrillic.
The folder where the project is located contains Polish characters. I changed the username, but path in the program remains the same. I don't know what else to do and I'm asking for help