I created a superclass called Figure that implements the Drawable interface. Following that, I created five subclasses of Figure, each with unique fields. Additionally, I overrode the draw method for each subclass.
I also developed a ColorSupplier class that provides a method to generate a random color string, chosen from a Color enum.

Finally, I implemented a FigureSupplier class, which contains a method that generates a Figure with random properties. In the end, I used a simple loop to generate and display a random selection of three figures, along with three default ones.